+++ src/rdkafka_conf.c
@@ -87,7 +87,14 @@ struct rd_kafka_property {
 	struct {
 		int val;
 		const char *str;
+                const char *unsupported; /**< Reason for value not being
+                                          *   supported in this build. */
 	} s2i[20];  /* _RK_C_S2I and _RK_C_S2F */
+
+        const char *unsupported; /**< Reason for propery not being supported
+                                  *   in this build.
+                                  *   Will be included in the conf_set()
+                                  *   error string. */
 
 	/* Value validator (STR) */
 	int (*validate) (const struct rd_kafka_property *prop,
@@ -111,6 +118,56 @@ struct rd_kafka_property {
 
 #define _RK(field)  offsetof(rd_kafka_conf_t, field)
 #define _RKT(field) offsetof(rd_kafka_topic_conf_t, field)
+
+#if WITH_SSL
+#define _UNSUPPORTED_SSL .unsupported = NULL
+#else
+#define _UNSUPPORTED_SSL .unsupported = "OpenSSL not available at build time"
+#endif
+
+#if OPENSSL_VERSION_NUMBER >= 0x1000200fL && defined(WITH_SSL) && !defined(LIBRESSL_VERSION_NUMBER)
+#define _UNSUPPORTED_OPENSSL_1_0_2 .unsupported = NULL
+#else
+#define _UNSUPPORTED_OPENSSL_1_0_2 .unsupported = \
+                "OpenSSL >= 1.0.2 not available at build time"
+#endif
+
+
+#if WITH_ZLIB
+#define _UNSUPPORTED_ZLIB .unsupported = NULL
+#else
+#define _UNSUPPORTED_ZLIB .unsupported = "zlib not available at build time"
+#endif
+
+#if WITH_SNAPPY
+#define _UNSUPPORTED_SNAPPY .unsupported = NULL
+#else
+#define _UNSUPPORTED_SNAPPY .unsupported = "snappy not enabled at build time"
+#endif
+
+#if WITH_ZSTD
+#define _UNSUPPORTED_ZSTD .unsupported = NULL
+#else
+#define _UNSUPPORTED_ZSTD .unsupported = "libzstd not available at build time"
+#endif
+
+#ifdef _WIN32
+#define _UNSUPPORTED_WIN32_GSSAPI .unsupported =                        \
+                "Kerberos keytabs are not supported on Windows, "       \
+                "instead the logged on "                                \
+                "user's credentials are used through native SSPI"
+#else
+        #define _UNSUPPORTED_WIN32_GSSAPI .unsupported = NULL
+#endif
+
+#if defined(_WIN32) || defined(WITH_SASL_CYRUS)
+#define _UNSUPPORTED_GSSAPI .unsupported = NULL
+#else
+#define _UNSUPPORTED_GSSAPI .unsupported = \
+                "cyrus-sasl/libsasl2 not available at build time"
+#endif
+
+#define _UNSUPPORTED_OAUTHBEARER _UNSUPPORTED_SSL
 
 
 static rd_kafka_conf_res_t
@@ -200,7 +257,7 @@ rd_kafka_conf_validate_broker_version (const struct rd_kafka_property *prop,
  */
 static RD_UNUSED int
 rd_kafka_conf_validate_single (const struct rd_kafka_property *prop,
-				const char *val, int ival) {
+                               const char *val, int ival) {
 	return !strchr(val, ',') && !strchr(val, ' ');
 }
 
@@ -230,37 +287,26 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
 	"An application can either query this value or attempt to set it "
 	"with its list of required features to check for library support.",
 	0, 0x7fffffff, 0xffff,
-	.s2i = {
-#if WITH_ZLIB
-		{ 0x1, "gzip" },
+          .s2i = {
+                        { 0x1, "gzip", _UNSUPPORTED_ZLIB },
+                        { 0x2, "snappy", _UNSUPPORTED_SNAPPY },
+                        { 0x4, "ssl", _UNSUPPORTED_SSL },
+                        { 0x8, "sasl" },
+                        { 0x10, "regex" },
+                        { 0x20, "lz4" },
+                        { 0x40, "sasl_gssapi", _UNSUPPORTED_GSSAPI },
+                        { 0x80, "sasl_plain" },
+                        { 0x100, "sasl_scram", _UNSUPPORTED_SSL },
+                        { 0x200, "plugins"
+#if !WITH_PLUGINS
+                          , .unsupported = "libdl/dlopen(3) not available at "
+                          "build time"
 #endif
-#if WITH_SNAPPY
-		{ 0x2, "snappy" },
-#endif
-#if WITH_SSL
-		{ 0x4, "ssl" },
-#endif
-                { 0x8, "sasl" },
-		{ 0x10, "regex" },
-		{ 0x20, "lz4" },
-#if defined(_WIN32) || WITH_SASL_CYRUS
-                { 0x40, "sasl_gssapi" },
-#endif
-                { 0x80, "sasl_plain" },
-#if WITH_SASL_SCRAM
-                { 0x100, "sasl_scram" },
-#endif
-#if WITH_PLUGINS
-                { 0x200, "plugins" },
-#endif
-#if WITH_ZSTD
-		{ 0x400, "zstd" },
-#endif
-#if WITH_SASL_OAUTHBEARER
-                { 0x800, "sasl_oauthbearer" },
-#endif
-		{ 0, NULL }
-		}
+                        },
+                        { 0x400, "zstd", _UNSUPPORTED_ZSTD },
+                        { 0x800, "sasl_oauthbearer", _UNSUPPORTED_SSL },
+                        { 0, NULL }
+                }
 	},
 	{ _RK_GLOBAL, "client.id", _RK_C_STR, _RK(client_id_str),
 	  "Client identifier.",
@@ -434,18 +480,22 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
 	  _RK(socket_rcvbuf_size),
 	  "Broker socket receive buffer size. System default is used if 0.",
 	  0, 100000000, 0 },
-#ifdef SO_KEEPALIVE
 	{ _RK_GLOBAL, "socket.keepalive.enable", _RK_C_BOOL,
 	  _RK(socket_keepalive),
           "Enable TCP keep-alives (SO_KEEPALIVE) on broker sockets",
-          0, 1, 0 },
+          0, 1, 0
+#ifndef SO_KEEPALIVE
+          , .unsupported = "SO_KEEPALIVE not available at build time"
 #endif
-#ifdef TCP_NODELAY
+        },
 	{ _RK_GLOBAL, "socket.nagle.disable", _RK_C_BOOL,
 	  _RK(socket_nagle_disable),
           "Disable the Nagle algorithm (TCP_NODELAY) on broker sockets.",
-          0, 1, 0 },
+          0, 1, 0
+#ifndef TCP_NODELAY
+          , .unsupported = "TCP_NODELAY not available at build time"
 #endif
+        },
         { _RK_GLOBAL, "socket.max.fails", _RK_C_INT,
           _RK(socket_max_fails),
           "Disconnect from broker when this number of send failures "
@@ -640,73 +690,77 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
 	  .vdef = RD_KAFKA_PROTO_PLAINTEXT,
 	  .s2i = {
 			{ RD_KAFKA_PROTO_PLAINTEXT, "plaintext" },
-#if WITH_SSL
-			{ RD_KAFKA_PROTO_SSL, "ssl" },
-#endif
+			{ RD_KAFKA_PROTO_SSL, "ssl", _UNSUPPORTED_SSL },
 			{ RD_KAFKA_PROTO_SASL_PLAINTEXT, "sasl_plaintext" },
-#if WITH_SSL
-			{ RD_KAFKA_PROTO_SASL_SSL, "sasl_ssl" },
-#endif
+			{ RD_KAFKA_PROTO_SASL_SSL, "sasl_ssl",
+                          _UNSUPPORTED_SSL },
 			{ 0, NULL }
 		} },
 
-#if WITH_SSL
 	{ _RK_GLOBAL, "ssl.cipher.suites", _RK_C_STR,
 	  _RK(ssl.cipher_suites),
 	  "A cipher suite is a named combination of authentication, "
 	  "encryption, MAC and key exchange algorithm used to negotiate the "
 	  "security settings for a network connection using TLS or SSL network "
 	  "protocol. See manual page for `ciphers(1)` and "
-	  "`SSL_CTX_set_cipher_list(3)."
+	  "`SSL_CTX_set_cipher_list(3).",
+          _UNSUPPORTED_SSL
 	},
-#if OPENSSL_VERSION_NUMBER >= 0x1000200fL && !defined(LIBRESSL_VERSION_NUMBER)
         { _RK_GLOBAL, "ssl.curves.list", _RK_C_STR,
           _RK(ssl.curves_list),
           "The supported-curves extension in the TLS ClientHello message specifies "
           "the curves (standard/named, or 'explicit' GF(2^k) or GF(p)) the client "
           "is willing to have the server use. See manual page for "
-          "`SSL_CTX_set1_curves_list(3)`. OpenSSL >= 1.0.2 required."
+          "`SSL_CTX_set1_curves_list(3)`. OpenSSL >= 1.0.2 required.",
+          _UNSUPPORTED_OPENSSL_1_0_2
         },
         { _RK_GLOBAL, "ssl.sigalgs.list", _RK_C_STR,
           _RK(ssl.sigalgs_list),
           "The client uses the TLS ClientHello signature_algorithms extension "
           "to indicate to the server which signature/hash algorithm pairs "
           "may be used in digital signatures. See manual page for "
-          "`SSL_CTX_set1_sigalgs_list(3)`. OpenSSL >= 1.0.2 required."
+          "`SSL_CTX_set1_sigalgs_list(3)`. OpenSSL >= 1.0.2 required.",
+          _UNSUPPORTED_OPENSSL_1_0_2
         },
-#endif
         { _RK_GLOBAL, "ssl.key.location", _RK_C_STR,
           _RK(ssl.key_location),
-          "Path to client's private key (PEM) used for authentication."
+          "Path to client's private key (PEM) used for authentication.",
+          _UNSUPPORTED_SSL
         },
         { _RK_GLOBAL|_RK_SENSITIVE, "ssl.key.password", _RK_C_STR,
           _RK(ssl.key_password),
           "Private key passphrase (for use with `ssl.key.location` "
-          "and `set_ssl_cert()`)"
+          "and `set_ssl_cert()`)",
+          _UNSUPPORTED_SSL
         },
         { _RK_GLOBAL|_RK_SENSITIVE, "ssl.key.pem", _RK_C_STR,
           _RK(ssl.key_pem),
-          "Client's private key string (PEM format) used for authentication."
+          "Client's private key string (PEM format) used for authentication.",
+          _UNSUPPORTED_SSL
         },
         { _RK_GLOBAL, "ssl_key", _RK_C_INTERNAL,
           _RK(ssl.key),
           "Client's private key as set by rd_kafka_conf_set_ssl_cert()",
           .dtor = rd_kafka_conf_cert_dtor,
-          .copy = rd_kafka_conf_cert_copy
+          .copy = rd_kafka_conf_cert_copy,
+          _UNSUPPORTED_SSL
         },
         { _RK_GLOBAL, "ssl.certificate.location", _RK_C_STR,
           _RK(ssl.cert_location),
-          "Path to client's public key (PEM) used for authentication."
+          "Path to client's public key (PEM) used for authentication.",
+          _UNSUPPORTED_SSL
         },
         { _RK_GLOBAL, "ssl.certificate.pem", _RK_C_STR,
           _RK(ssl.cert_pem),
-          "Client's public key string (PEM format) used for authentication."
+          "Client's public key string (PEM format) used for authentication.",
+          _UNSUPPORTED_SSL
         },
         { _RK_GLOBAL, "ssl_certificate", _RK_C_INTERNAL,
           _RK(ssl.key),
           "Client's public key as set by rd_kafka_conf_set_ssl_cert()",
           .dtor = rd_kafka_conf_cert_dtor,
-          .copy = rd_kafka_conf_cert_copy
+          .copy = rd_kafka_conf_cert_copy,
+          _UNSUPPORTED_SSL
         },
 
         { _RK_GLOBAL, "ssl.ca.location", _RK_C_STR,
@@ -723,34 +777,39 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
           "`probe` a list of standard paths will be probed and the first one "
           "found will be used as the default CA certificate location path. "
           "If OpenSSL is dynamically linked the OpenSSL library's default "
-          "path will be used (see `OPENSSLDIR` in `openssl version -a`)."
+          "path will be used (see `OPENSSLDIR` in `openssl version -a`).",
+          _UNSUPPORTED_SSL
         },
         { _RK_GLOBAL, "ssl_ca", _RK_C_INTERNAL,
           _RK(ssl.ca),
           "CA certificate as set by rd_kafka_conf_set_ssl_cert()",
           .dtor = rd_kafka_conf_cert_dtor,
-          .copy = rd_kafka_conf_cert_copy
+          .copy = rd_kafka_conf_cert_copy,
+          _UNSUPPORTED_SSL
         },
         { _RK_GLOBAL, "ssl.crl.location", _RK_C_STR,
           _RK(ssl.crl_location),
-          "Path to CRL for verifying broker's certificate validity."
+          "Path to CRL for verifying broker's certificate validity.",
+          _UNSUPPORTED_SSL
         },
         { _RK_GLOBAL, "ssl.keystore.location", _RK_C_STR,
           _RK(ssl.keystore_location),
-          "Path to client's keystore (PKCS#12) used for authentication."
+          "Path to client's keystore (PKCS#12) used for authentication.",
+          _UNSUPPORTED_SSL
         },
         { _RK_GLOBAL|_RK_SENSITIVE, "ssl.keystore.password", _RK_C_STR,
           _RK(ssl.keystore_password),
-          "Client's keystore (PKCS#12) password."
+          "Client's keystore (PKCS#12) password.",
+          _UNSUPPORTED_SSL
         },
         { _RK_GLOBAL, "enable.ssl.certificate.verification", _RK_C_BOOL,
           _RK(ssl.enable_verify),
           "Enable OpenSSL's builtin broker (server) certificate verification. "
           "This verification can be extended by the application by "
           "implementing a certificate_verify_cb.",
-          0, 1, 1
+          0, 1, 1,
+          _UNSUPPORTED_SSL
         },
-#if OPENSSL_VERSION_NUMBER >= 0x1000200fL
         { _RK_GLOBAL, "ssl.endpoint.identification.algorithm", _RK_C_S2I,
           _RK(ssl.endpoint_identification),
           "Endpoint identification algorithm to validate broker "
@@ -763,14 +822,14 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
           .s2i = {
                         { RD_KAFKA_SSL_ENDPOINT_ID_NONE, "none" },
                         { RD_KAFKA_SSL_ENDPOINT_ID_HTTPS, "https" }
-                }
+                },
+          _UNSUPPORTED_OPENSSL_1_0_2
         },
-#endif
         { _RK_GLOBAL, "ssl.certificate.verify_cb", _RK_C_PTR,
           _RK(ssl.cert_verify_cb),
-          "Callback to verify the broker certificate chain."
+          "Callback to verify the broker certificate chain.",
+          _UNSUPPORTED_SSL
         },
-#endif /* WITH_SSL */
 
         /* Point user in the right direction if they try to apply
          * Java client SSL / JAAS properties. */
@@ -778,7 +837,8 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
           _RK(dummy),
           "Java TrustStores are not supported, use `ssl.ca.location` "
           "and a certificate file instead. "
-          "See https://github.com/edenhill/librdkafka/wiki/Using-SSL-with-librdkafka for more information."
+          "See https://github.com/edenhill/librdkafka/wiki/Using-SSL-with-librdkafka "
+          "for more information."
         },
         { _RK_GLOBAL, "sasl.jaas.config", _RK_C_INVALID,
           _RK(dummy),
@@ -806,7 +866,6 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
           "This client's Kerberos principal name. "
           "(Not supported on Windows, will use the logon user's principal).",
 	  .sdef = "kafkaclient" },
-#ifndef _WIN32
         { _RK_GLOBAL, "sasl.kerberos.kinit.cmd", _RK_C_STR,
           _RK(sasl.kinit_cmd),
           "Shell command to refresh or acquire the client's Kerberos ticket. "
@@ -818,27 +877,30 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
           /* First attempt to refresh, else acquire. */
           "kinit -R -t \"%{sasl.kerberos.keytab}\" "
           "-k %{sasl.kerberos.principal} || "
-          "kinit -t \"%{sasl.kerberos.keytab}\" -k %{sasl.kerberos.principal}"
+          "kinit -t \"%{sasl.kerberos.keytab}\" -k %{sasl.kerberos.principal}",
+          _UNSUPPORTED_WIN32_GSSAPI
         },
         { _RK_GLOBAL, "sasl.kerberos.keytab", _RK_C_STR,
           _RK(sasl.keytab),
           "Path to Kerberos keytab file. "
           "This configuration property is only used as a variable in "
           "`sasl.kerberos.kinit.cmd` as "
-          "` ... -t \"%{sasl.kerberos.keytab}\"`." },
+          "` ... -t \"%{sasl.kerberos.keytab}\"`.",
+          _UNSUPPORTED_WIN32_GSSAPI
+        },
         { _RK_GLOBAL, "sasl.kerberos.min.time.before.relogin", _RK_C_INT,
           _RK(sasl.relogin_min_time),
           "Minimum time in milliseconds between key refresh attempts. "
           "Disable automatic key refresh by setting this property to 0.",
-          0, 86400*1000, 60*1000 },
-#endif
+          0, 86400*1000, 60*1000,
+          _UNSUPPORTED_WIN32_GSSAPI
+        },
 	{ _RK_GLOBAL|_RK_HIGH, "sasl.username", _RK_C_STR,
 	  _RK(sasl.username),
 	  "SASL username for use with the PLAIN and SASL-SCRAM-.. mechanisms" },
 	{ _RK_GLOBAL|_RK_HIGH, "sasl.password", _RK_C_STR,
 	  _RK(sasl.password),
 	  "SASL password for use with the PLAIN and SASL-SCRAM-.. mechanism" },
-#if WITH_SASL_OAUTHBEARER
         { _RK_GLOBAL, "sasl.oauthbearer.config", _RK_C_STR,
           _RK(sasl.oauthbearer_config),
           "SASL/OAUTHBEARER configuration. The format is "
@@ -855,32 +917,40 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
           "scope=role1,role2 lifeSeconds=600`. In addition, SASL extensions "
           "can be communicated to the broker via "
           "`extension_NAME=value`. For example: "
-          "`principal=admin extension_traceId=123`" },
+          "`principal=admin extension_traceId=123`",
+          _UNSUPPORTED_OAUTHBEARER
+        },
         { _RK_GLOBAL, "enable.sasl.oauthbearer.unsecure.jwt", _RK_C_BOOL,
           _RK(sasl.enable_oauthbearer_unsecure_jwt),
           "Enable the builtin unsecure JWT OAUTHBEARER token handler "
           "if no oauthbearer_refresh_cb has been set. "
           "This builtin handler should only be used for development "
           "or testing, and not in production.",
-          0, 1, 0 },
+          0, 1, 0,
+          _UNSUPPORTED_OAUTHBEARER
+        },
         { _RK_GLOBAL, "oauthbearer_token_refresh_cb", _RK_C_PTR,
           _RK(sasl.oauthbearer_token_refresh_cb),
           "SASL/OAUTHBEARER token refresh callback (set with "
           "rd_kafka_conf_set_oauthbearer_token_refresh_cb(), triggered by "
           "rd_kafka_poll(), et.al. "
           "This callback will be triggered when it is time to refresh "
-          "the client's OAUTHBEARER token." },
-#endif
+          "the client's OAUTHBEARER token.",
+          _UNSUPPORTED_OAUTHBEARER
+        },
 
-#if WITH_PLUGINS
         /* Plugins */
         { _RK_GLOBAL, "plugin.library.paths", _RK_C_STR,
           _RK(plugin_paths),
           "List of plugin libraries to load (; separated). "
           "The library search path is platform dependent (see dlopen(3) for Unix and LoadLibrary() for Windows). If no filename extension is specified the "
           "platform-specific extension (such as .dll or .so) will be appended automatically.",
-          .set = rd_kafka_plugins_conf_set },
+#if WITH_PLUGINS
+          .set = rd_kafka_plugins_conf_set
+#else
+          .unsupported = "libdl/dlopen(3) not available at build time"
 #endif
+        },
 
         /* Interceptors are added through specific API and not exposed
          * as configuration properties.
@@ -1242,18 +1312,16 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
 	  .vdef = RD_KAFKA_COMPRESSION_NONE,
 	  .s2i = {
 			{ RD_KAFKA_COMPRESSION_NONE,   "none" },
-#if WITH_ZLIB
-			{ RD_KAFKA_COMPRESSION_GZIP,   "gzip" },
-#endif
-#if WITH_SNAPPY
-			{ RD_KAFKA_COMPRESSION_SNAPPY, "snappy" },
-#endif
+			{ RD_KAFKA_COMPRESSION_GZIP,   "gzip",
+                          _UNSUPPORTED_ZLIB },
+			{ RD_KAFKA_COMPRESSION_SNAPPY, "snappy",
+                          _UNSUPPORTED_SNAPPY },
 			{ RD_KAFKA_COMPRESSION_LZ4, "lz4" },
-#if WITH_ZSTD
-			{ RD_KAFKA_COMPRESSION_ZSTD, "zstd" },
-#endif
+			{ RD_KAFKA_COMPRESSION_ZSTD, "zstd",
+                          _UNSUPPORTED_ZSTD },
 			{ 0 }
-		} },
+		}
+        },
         { _RK_GLOBAL|_RK_PRODUCER|_RK_MED, "compression.type", _RK_C_ALIAS,
           .sdef = "compression.codec" },
         { _RK_GLOBAL|_RK_PRODUCER|_RK_MED, "batch.num.messages", _RK_C_INT,
@@ -1377,19 +1445,17 @@ static const struct rd_kafka_property rd_kafka_properties[] = {
 	  .vdef = RD_KAFKA_COMPRESSION_INHERIT,
 	  .s2i = {
 		  { RD_KAFKA_COMPRESSION_NONE, "none" },
-#if WITH_ZLIB
-		  { RD_KAFKA_COMPRESSION_GZIP, "gzip" },
-#endif
-#if WITH_SNAPPY
-		  { RD_KAFKA_COMPRESSION_SNAPPY, "snappy" },
-#endif
+		  { RD_KAFKA_COMPRESSION_GZIP, "gzip",
+                    _UNSUPPORTED_ZLIB },
+		  { RD_KAFKA_COMPRESSION_SNAPPY, "snappy",
+                    _UNSUPPORTED_SNAPPY },
 		  { RD_KAFKA_COMPRESSION_LZ4, "lz4" },
-#if WITH_ZSTD
-		  { RD_KAFKA_COMPRESSION_ZSTD, "zstd" },
-#endif
+		  { RD_KAFKA_COMPRESSION_ZSTD, "zstd",
+                    _UNSUPPORTED_ZSTD },
 		  { RD_KAFKA_COMPRESSION_INHERIT, "inherit" },
 		  { 0 }
-		} },
+		}
+        },
         { _RK_TOPIC|_RK_PRODUCER|_RK_HIGH, "compression.type", _RK_C_ALIAS,
           .sdef = "compression.codec" },
         { _RK_TOPIC|_RK_PRODUCER|_RK_MED, "compression.level", _RK_C_INT,
@@ -1731,6 +1797,14 @@ rd_kafka_anyconf_set_prop (int scope, void *conf,
 			   char *errstr, size_t errstr_size) {
 	int ival;
 
+        if (prop->unsupported) {
+                rd_snprintf(errstr, errstr_size,
+                            "Configuration property \"%s\" not supported "
+                            "in this build: %s",
+                            prop->name, prop->unsupported);
+                return RD_KAFKA_CONF_INVALID;
+        }
+
 	switch (prop->type)
 	{
 	case _RK_C_STR:
@@ -1833,6 +1907,15 @@ rd_kafka_anyconf_set_prop (int scope, void *conf,
 					    prop->name);
 				return RD_KAFKA_CONF_INVALID;
 			}
+
+                        if (prop->s2i[match].unsupported) {
+                                rd_snprintf(errstr, errstr_size,
+                                            "Unsupported value \"%s\" for "
+                                            "configuration property \"%s\": %s",
+                                            value, prop->name,
+                                            prop->s2i[match].unsupported);
+                                return RD_KAFKA_CONF_INVALID;
+                        }
 
 			ival = prop->s2i[match].val;
 		}
@@ -1961,6 +2044,17 @@ rd_kafka_anyconf_set_prop (int scope, void *conf,
 					new_val = prop->s2i[j].val;
 				else
 					continue;
+
+                                if (prop->s2i[j].unsupported) {
+                                        rd_snprintf(
+                                                errstr, errstr_size,
+                                                "Unsupported value \"%.*s\" "
+                                                "for configuration property "
+                                                "\"%s\": %s",
+                                                (int)(t-s), s, prop->name,
+                                                prop->s2i[j].unsupported);
+                                        return RD_KAFKA_CONF_INVALID;
+                                }
 
 				rd_kafka_anyconf_set_prop0(scope, conf, prop,
                                                            value, new_val,
@@ -2724,13 +2818,18 @@ void rd_kafka_topic_conf_set_opaque (rd_kafka_topic_conf_t *topic_conf,
  *
  * An \p ival of -1 means all.
  *
+ * @param include_unsupported Include flag values that are unsupported
+ *                            due to missing dependencies at build time.
+ *
  * @returns the number of bytes written to \p dest (if not NULL), else the
  *          total number of bytes needed.
  *
  */
+static
 size_t rd_kafka_conf_flags2str (char *dest, size_t dest_size, const char *delim,
 				const struct rd_kafka_property *prop,
-				int ival) {
+				int ival,
+                                rd_bool_t include_unsupported) {
 	size_t of = 0;
 	int j;
 
@@ -2746,6 +2845,8 @@ size_t rd_kafka_conf_flags2str (char *dest, size_t dest_size, const char *delim,
 		else if (prop->type == _RK_C_S2I &&
 			   ival != -1 && prop->s2i[j].val != ival)
 			continue;
+                else if (prop->s2i[j].unsupported && !include_unsupported)
+                        continue;
 
 		if (!dest)
 			of += strlen(prop->s2i[j].str) + (of > 0 ? 1 : 0);
@@ -2833,7 +2934,8 @@ rd_kafka_anyconf_get0 (const void *conf, const struct rd_kafka_property *prop,
 
 		val_len = rd_kafka_conf_flags2str(dest,
                                                   dest ? *dest_size : 0, ",",
-						  prop, ival);
+						  prop, ival,
+                                                  rd_false/*only supported*/);
 		if (dest) {
 			val_len = 0;
 			val = dest;
@@ -3047,8 +3149,10 @@ void rd_kafka_conf_properties_show (FILE *fp) {
 			if (prop->type == _RK_C_PATLIST)
 				typeinfo = "pattern list";
 			if (prop->s2i[0].str) {
-				rd_kafka_conf_flags2str(tmp, sizeof(tmp), ", ",
-							prop, -1);
+				rd_kafka_conf_flags2str(
+                                        tmp, sizeof(tmp), ", ",
+                                        prop, -1,
+                                        rd_true/*include unsupported*/);
 				fprintf(fp, "%-15s | %13s",
 					tmp, prop->sdef ? prop->sdef : "");
 			} else {
@@ -3076,7 +3180,8 @@ void rd_kafka_conf_properties_show (FILE *fp) {
 		case _RK_C_S2I:
 			typeinfo = "enum value";
 			rd_kafka_conf_flags2str(tmp, sizeof(tmp), ", ",
-						prop, -1);
+						prop, -1,
+                                                rd_true/*include unsupported*/);
 			fprintf(fp, "%-15s | ", tmp);
 
 			for (j = 0 ; j < (int)RD_ARRAYSIZE(prop->s2i); j++) {
@@ -3096,11 +3201,14 @@ void rd_kafka_conf_properties_show (FILE *fp) {
 			if (!strcmp(prop->name, "builtin.features"))
 				*tmp = '\0';
 			else
-				rd_kafka_conf_flags2str(tmp, sizeof(tmp), ", ",
-							prop, -1);
+				rd_kafka_conf_flags2str(
+                                        tmp, sizeof(tmp), ", ",
+                                        prop, -1,
+                                        rd_true/*include unsupported*/);
 			fprintf(fp, "%-15s | ", tmp);
 			rd_kafka_conf_flags2str(tmp, sizeof(tmp), ", ",
-						prop, prop->vdef);
+						prop, prop->vdef,
+                                                rd_true/*include unsupported*/);
 			fprintf(fp, "%13s", tmp);
 
 			break;